Refactor avatar logic
Mainly to bring the file down to a sane size
