+++ kano_avatar/logic.py
@@ -6,997 +6,19 @@
 # Logic for parsing and creating avatars for a Kano World profile
 import os
 import random
-from PIL import Image
 
-from kano_avatar.paths import (AVATAR_CONF_FILE, CHARACTER_DIR, ITEM_DIR,
-                               ACTIVE_CATEGORY_ICONS, INACTIVE_CATEGORY_ICONS,
-                               ACTIVE_SPECIAL_CATEGORY_ICONS,
-                               INACTIVE_SPECIAL_CATEGORY_ICONS,
-                               CIRC_ASSET_MASK, RING_ASSET, PREVIEW_ICONS,
-                               ENVIRONMENT_DIR, AVATAR_SCRATCH,
-                               AVATAR_DEFAULT_LOC, AVATAR_DEFAULT_NAME,
-                               PLAIN_MASK, AVATAR_SELECTED_ITEMS)
+from kano_avatar.paths import (AVATAR_CONF_FILE,
+                               AVATAR_SCRATCH, AVATAR_DEFAULT_LOC,
+                               AVATAR_DEFAULT_NAME, AVATAR_SELECTED_ITEMS)
 from kano.logging import logger
 
-from kano_profile.badges import calculate_badges
 from kano_profile.profile import (set_avatar, set_environment,
                                   save_profile_variable)
 from kano.utils import get_date_now
 from json import dump, load
 
+from .conf_parser import AvatarConfParser
 # TODO Check which types of names are case sensitive
-
-
-class AvatarBaseAccessory:
-    _name = ''
-    _asset_fname = ''
-    _img_preview = ''
-    _date_created = ''
-    _display_order = 0
-    _is_unlocked = False
-    _img = None
-    _unique_id = ''
-
-    def name(self):
-        """ Provides the display name of the Item
-        :returns: display name of character as a string
-        :rtype: string
-        """
-        return self._name
-
-    def is_unlocked(self):
-        """ Returns the locked state of the environment
-        :returns: lock state
-        :rtype: Boolean
-        """
-        return self._is_unlocked
-
-    def get_preview_img(self):
-        """ Provides the Background preview image path
-        :returns: absolute path to preview image as a string
-        :rtype: string
-        """
-        return self._img_preview
-
-    def get_date(self):
-        """ Provides the creation date of the environment
-        :returns: creation date as a string
-        :rtype: string
-        """
-        return self._date_created
-
-    def get_disp_order(self):
-        """ Provides the display order of the environment
-        :returns: display index as an integer
-        :rtype: integer
-        """
-        return self._display_order
-
-    def get_fname(self):
-        """ Provides the item's asset filename
-        :returns: filename as a string
-        :rtype: string
-        """
-        return self._asset_fname
-
-    def get_id(self):
-        """ Provides the unique id for the item
-        :returns: unique id as a string
-        :rtype: string
-        """
-        return self._unique_id
-
-
-class AvatarAccessory(AvatarBaseAccessory):
-    """ Class for handling items and applying them onto an Avatar Character
-    class.
-    """
-    _category = ''
-    _img_position_x = 0
-    _img_position_y = 0
-
-    def __init__(self, name, category, file_name, preview_img, x, y,
-                 date_created, item_id, is_unlocked, display_order):
-        self._category = category
-        self._name = name
-        self._img_position_x = x
-        self._img_position_y = y
-        self._date_created = date_created
-        self._unique_id = item_id
-        self._display_order = display_order
-        self._is_unlocked = is_unlocked
-        # if an absolute path is given use it instead, so that we can
-        # override elements
-        if os.path.isabs(file_name):
-            self._asset_fname = file_name
-        else:
-            self._asset_fname = os.path.join(ITEM_DIR, file_name)
-
-        if os.path.isabs(preview_img):
-            self._img_preview = preview_img
-        else:
-            self._img_preview = os.path.join(PREVIEW_ICONS, preview_img)
-
-    def category(self):
-        """ Provides the category name to which the Item belongs to
-        :returns: category name as a string
-        :rtype: string
-        """
-        return self._category
-
-    def load_image(self):
-        """ Loads the asset's image internally. This is necessary before
-        pasting the item over a character, however there is no need to call it
-        externally.
-        """
-        self._img = Image.open(self.get_fname())
-
-    def paste_over_image(self, img):
-        """ Pastes the item's image over a character's base image. The
-        position of where the image will be pasted is specified by the x,y
-        coordinates given during the class' instantiation. (x,y are distances
-        in pixels of the top left corner)
-        :param img: PIL Image class on which the pasting will occur
-        """
-        if self._img is None:
-            self.load_image()
-        # position of upper left corner
-        position = (self._img_position_x, self._img_position_y)
-        # We need to remove the transparency of the item so that it doesn't
-        # cause a gap on the base layer.
-        r, g, b, a = self._img.split()
-        item = Image.merge('RGB', (r, g, b))
-        transp_mask = Image.merge("L", (a,))
-        img.paste(item, position, transp_mask)
-
-
-class AvatarCharacter(AvatarBaseAccessory):
-    """ Class for handling an Avatar character. It holds the image data for
-    the character so as to serve as a base on which the items are pasted on.
-    """
-    _crop_x = 0
-    _crop_y = 0
-
-    def __init__(self, name, file_name, preview_img, x, y, date_created,
-                 char_id, is_unlocked, display_order):
-        self._name = name
-        if os.path.isabs(file_name):
-            self._asset_fname = file_name
-        else:
-            self._asset_fname = os.path.join(CHARACTER_DIR, file_name)
-
-        if os.path.isabs(preview_img):
-            self._img_preview = preview_img
-        else:
-            self._img_preview = os.path.join(CHARACTER_DIR, preview_img)
-        self._crop_x = x
-        self._crop_y = y
-        self._display_order = display_order
-        self._unique_id = char_id
-        self._date_created = date_created
-        self._is_unlocked = is_unlocked
-
-    def load_image(self):
-        """ Loads the character's image internally. This is necessary before
-        pasting the item over a character.
-        """
-        self._img = Image.open(self._asset_fname)
-
-    def get_img(self):
-        """ Get the image class for the character.
-        :returns: Image class (from PIL module)
-        :rtype: Image class (from PIL module)
-        """
-        return self._img
-
-    def save_image(self, file_name):
-        """ Save character image (together with items that have been pasted on
-        it), to a file.
-        :param file_name: filename to be saved to as a string
-        """
-        self._img.save(file_name)
-
-    def generate_circular_assets(self, file_name_plain, file_name_ring):
-        """ This function creates the circular assets that are required for the
-        kano profile
-        :param file_name: Path to where the completed file should be saved as a
-                          string
-        """
-        rc_plain = self._generate_plain_circular(file_name_plain)
-
-        # 54 is the size of the stamp icon we require
-        rc_ring = self._generate_white_circular(file_name_ring, resize=54)
-        return rc_plain and rc_ring
-
-    def _generate_white_circular(self, file_name, resize=0):
-        return self._generate_ring_circular(
-            file_name, RING_ASSET, CIRC_ASSET_MASK, resize=resize)
-
-    def _generate_plain_circular(self, file_name):
-        return self._generate_ring_circular(
-            file_name, PLAIN_MASK, PLAIN_MASK, invert_order=True)
-
-    def _generate_ring_circular(self, file_name, ring, mask,
-                                invert_order=False, resize=0):
-        """ Creates the circular asset to be used in the top left corner on the
-        desktop. The mask and ring assets need to have the same size.
-        :param file_name: Path to where the completed (and resized if set so)
-                          file will be saved as a string
-        :param ring: Path to the ring asset to be used
-        :param mask: Path to the mast asset to be used
-        :param invert_order: invert order of the ring and mask assets in the
-                             composite function
-        :type invert_order: Boolean
-        :param resize: (Optional) The size of the side of the resulting resized
-                       image (Will be a square)
-        :type resize: integer
-        :returns: True iff no errors have occurred
-        :rtype: Boolean
-        """
-        ring_img = Image.open(ring)
-        mask_img = Image.open(mask)
-
-        if ring_img.size != mask_img.size:
-            logger.warn('Mask and ring asset do not have the same size')
-            return False
-
-        cropped_img = self._get_image_cropped(
-            mask_img.size[0], mask_img.size[1])
-
-        if not invert_order:
-            img_out = Image.composite(ring_img, cropped_img, mask_img)
-        else:
-            img_out = Image.composite(cropped_img, ring_img, mask_img)
-
-        # Resize to specified width if set
-        if resize != 0:
-            if img_out.size[0] != img_out.size[1]:
-                logger.warn("Image is not square, resizing it will distort it")
-            if resize < 0:
-                logger.error(
-                    "Resize value negative: {}, won't continue".format(resize)
-                )
-                return False
-            img_out.thumbnail((resize, resize), Image.ANTIALIAS)
-
-        img_out.save(file_name)
-        logger.debug("created {}".format(file_name))
-        return True
-
-    def _get_image_cropped(self, size_x, size_y):
-        """ Returns an instance of an image cropped to the specified
-        dimensions, centered at the coordinates detailed by the class
-        variables _crop_x and _crop_y.
-        :param size_x: Size of the final image in the x dimension
-        :param size_y: Size of the final image in the y dimension
-        :returns: Instance of PIL Image
-        :rtype: Image class from PIL module
-        """
-        # Create a box of useful image data
-        x_left = int(size_x / 2)
-        x_right = size_x - x_left
-
-        y_up = int(size_y / 2)
-        y_down = size_y - y_up
-
-        box = (self._crop_x - x_left,
-               self._crop_y - y_up,
-               self._crop_x + x_right,
-               self._crop_y + y_down)
-
-        return self._img.crop(box)
-
-
-class AvatarEnvironment(AvatarBaseAccessory):
-    """ Class for handling the environment (background) for a character. As
-    it contains the image that will work as the background (lowest z-index)
-    but also the largest in terms of size, it deserves a class of its own.
-    """
-
-    def __init__(self, name, file_name, preview_img, date_created, env_id,
-                 is_unlocked, display_order):
-        self._name = name
-        self._date_created = date_created
-        self._display_order = display_order
-        self._unique_id = env_id
-        self._is_unlocked = is_unlocked
-        if os.path.isabs(file_name):
-            self._asset_fname = file_name
-        else:
-            self._asset_fname = os.path.join(ENVIRONMENT_DIR, file_name)
-
-        if os.path.isabs(preview_img):
-            self._img_preview = preview_img
-        else:
-            self._img_preview = os.path.join(PREVIEW_ICONS, preview_img)
-
-    def load_image(self):
-        """ Loads the environment image internally.
-        """
-        self._img = Image.open(self._asset_fname).convert('RGBA')
-
-    def attach_char(self, char_img, x=0.5, y=0.5):
-        """ Attach a character to the environment. This method will overwrite
-        the internal image to include the character with the environment in
-        the background. In order to reset the environment you may use the
-        .load_image() method
-        :param char_img: Image object to be attached to the environment
-        :param x: Value between [0,1) to control the position of the character
-                  on the environment. It can be thought of like a percentage
-                  of the total x axis of the image where the centre of the new
-                  image will be pasted
-        :param y: Value between [0,1) to control the position of the character
-                  on the environment. It can be thought of like a percentage
-                  of the total y axis of the image where the centre of the new
-                  image will be pasted
-        :returns: False if the operation is not successful (with appropriate
-                  logging)
-        :rtype: Boolean
-        """
-
-        if not char_img:
-            logger.warn(
-                "Char image to be attached to the env is None, will exit")
-            return False
-
-        if x < 0 or x >= 1:
-            err_msg = 'Argument x given to attach_char is out of bounds [0,1)'
-            logger.error(err_msg)
-            return False
-        if y < 0 or y >= 1:
-            err_msg = 'Argument x given to attach_char is out of bounds [0,1)'
-            logger.error(err_msg)
-            return False
-
-        if not self._img:
-            err_msg = "Internal envir Image hasn't been loaded. Will load now"
-            logger.debug(err_msg)
-            self.load_image()
-
-        # Resize avatar if we can't fit it in (Normally shouldn't happen
-        # but it makes testing without properly sized assets easier)
-        char_szx = char_img.size[0]
-        char_szy = char_img.size[1]
-
-        if char_szx > self.get_img().size[0] or \
-           char_szy > self.get_img().size[1]:
-            # First calculate the reduction coefficient
-            # so that no side is larger than 90% of the background canvas
-            coeff = 0.9
-            c_x = coeff * self.get_img().size[0] / char_szx
-            c_y = coeff * self.get_img().size[1] / char_szy
-            c = min(c_x, c_y)
-            new_size = (int(c * char_szx), int(c * char_szy))
-            resized_char = char_img.resize(new_size, Image.ANTIALIAS)
-        else:
-            resized_char = char_img
-
-        top_left_x = int(self.get_img().size[0] * x - resized_char.size[0]/2)
-        top_left_y = int(self.get_img().size[1] * y - resized_char.size[1]/2)
-
-        # We need to create a temporary image to hold the avatar since
-        # in order to composite images they need to have the same size
-        temp_img = Image.new(self.get_img().mode, self.get_img().size)
-        temp_img.paste(resized_char, (top_left_x, top_left_y))
-
-        self._img = Image.composite(temp_img, self.get_img(), temp_img)
-
-        return True
-
-    def get_img(self):
-        """ Get the image class for the character.
-        :returns: the internally used Image class (from PIL module)
-        :rtype: Image class (from PIL module)
-        """
-        return self._img
-
-    def save_image(self, file_name):
-        """ Save character image (together with items that have been pasted on
-        it), to a file.
-        :param file_name: filename to be saved to as a string
-        :returns: Always True (for now, wink wink)
-        :rtype: Boolean
-        """
-        # TODO Error checking
-        self._img.save(file_name)
-        return True
-
-
-class AvatarConfParser:
-    """ A class to take on the important task of parsing the configuration
-    file used for generating new Avatars.
-    Please use this class only if you require only parsing of the
-    conf file, otherwise please use the AvatarCreator() class
-    """
-    _categories = set()
-    _zindex = set()
-    _cat_to_z_index = {}
-    _cat_to_disp_order = {}
-    _zindex_to_categories = {}
-    _objects = {}
-    _characters = {}
-    _environments = {}
-    _object_per_cat = {}
-    _inactive_category_icons = {}
-    _active_category_icons = {}
-    _selected_borders = {}
-    _hover_borders = {}
-    _inactive_special_category_icons = {}
-    _active_special_category_icons = {}
-    _border_special_cat = {}
-    _hover_border_special_cat = {}
-    categories_label = 'categories'
-    objects_label = 'objects'
-    char_label = 'characters'
-    env_label = 'environments'
-    spec_cat_label = 'special_categories'
-    special_category_labels = [char_label, env_label]
-    _special_cat_to_disp_order = {}
-
-    def __init__(self, conf_data):
-        if self.categories_label not in conf_data:
-            logger.error('{} dict not found'.format(self.categories_label))
-        else:
-            self._populate_cat_structures(conf_data[self.categories_label])
-
-        if self.char_label not in conf_data:
-            logger.error('{} dict not found'.format(self.char_label))
-        else:
-            self._populate_character_structures(conf_data)
-
-        if self.objects_label not in conf_data:
-            logger.error('{} dict not found'.format(self.objects_label))
-        else:
-            self._populate_object_structures(conf_data)
-
-        self._populate_environment_structures(conf_data)
-
-        if self.spec_cat_label not in conf_data:
-            logger.error('{} dict not found'.format(self.spec_cat_label))
-        else:
-            self._populate_special_category_structures(conf_data)
-
-    def _populate_cat_structures(self, categories):
-        """ Populates internal structures related to categories
-        :param categories: categories section of YAML format configuration
-                           structure read from file
-        """
-        for cat in categories:
-            self._cat_to_z_index[cat['cat_name']] = cat['z_index']
-            icon_file = cat['disp_icon']
-            # TODO Fix the following if the icon_file is absolute
-            if not os.path.isabs(icon_file):
-                active_icon_file = os.path.join(ACTIVE_CATEGORY_ICONS,
-                                                icon_file)
-                inactive_icon_file = os.path.join(INACTIVE_CATEGORY_ICONS,
-                                                  icon_file)
-            self._active_category_icons[cat['cat_name']] = active_icon_file
-            self._inactive_category_icons[cat['cat_name']] = inactive_icon_file
-
-            selected_border_file = os.path.join(PREVIEW_ICONS,
-                                                cat['selected_border'])
-            self._selected_borders[cat['cat_name']] = selected_border_file
-            hover_border_file = os.path.join(PREVIEW_ICONS,
-                                             cat['hover_border'])
-            self._hover_borders[cat['cat_name']] = hover_border_file
-            self._cat_to_disp_order[cat['cat_name']] = cat['display_order']
-
-        # Save both the unique set of z-indexes and categories
-        self._categories = set(self._cat_to_z_index.keys())
-        self._zindex = set(self._cat_to_z_index.values())
-
-        self._zindex_to_categories = {i: [] for i in self._zindex}
-        for cat_name, zindex in self._cat_to_z_index.items():
-            self._zindex_to_categories[zindex].append(cat_name)
-
-    def _populate_object_structures(self, conf_data):
-        """ Populates internal structures related to items
-        :param conf_data: YAML format configuration structure read from file
-        """
-        for obj in conf_data[self.objects_label]:
-            new_name = obj['display_name']
-            new_cat = obj['category']
-            new_fname = obj['img_name']
-            new_prev_img = obj['preview_img']
-            new_x = obj['position_x']
-            new_y = obj['position_y']
-            new_disp_ord = obj['display_order']
-            new_date = obj['date_created']
-            new_id = obj['item_id']
-            # TODO hardcoded all unlocked for items until we change them to
-            # use the same scheme as environments
-            new_unlock = True
-            new_obj = AvatarAccessory(new_name,
-                                      new_cat,
-                                      new_fname,
-                                      new_prev_img,
-                                      new_x,
-                                      new_y,
-                                      new_date,
-                                      new_id,
-                                      new_unlock,
-                                      new_disp_ord)
-            self._objects[new_name] = new_obj
-            if new_cat not in self._object_per_cat:
-                self._object_per_cat[new_cat] = []
-            self._object_per_cat[new_cat].append(new_obj)
-
-    def _populate_character_structures(self, conf_data):
-        """ Populates internal structures related to characters
-        :param conf_data: YAML format configuration structure read from file
-        """
-        for obj in conf_data[self.char_label]:
-            new_name = obj['display_name']
-            new_fname = obj['img_name']
-            new_prev_img = obj['preview_img']
-            x = obj['crop_x']
-            y = obj['crop_y']
-            new_disp_ord = obj['display_order']
-            new_date = obj['date_created']
-            new_id = obj['character_id']
-            # TODO hardcoded all unlocked for characters until we change them
-            # to use the same scheme as environments
-            new_unlock = True
-            new_obj = AvatarCharacter(new_name,
-                                      new_fname,
-                                      new_prev_img,
-                                      x,
-                                      y,
-                                      new_date,
-                                      new_id,
-                                      new_unlock,
-                                      new_disp_ord)
-            self._characters[new_name] = new_obj
-
-    def _populate_environment_structures(self, conf_data):
-        """ Populates internal structures related to environments (backgrounds)
-        :param conf_data: YAML format configuration structure read from file
-        """
-        # TODO conf_data is not used any more in this function, maybe remove it
-        envirs = calculate_badges()
-
-        for _, env in envirs[self.env_label]['all'].iteritems():
-            new_name = env['title']
-            new_fname = env['img_name']
-            new_id = env['item_id']
-            new_disp_ord = env['display_order']
-            new_date = env['date_created']
-            new_unlocked = env['achieved']
-            if new_unlocked:
-                new_prev_img = env['preview_img']
-            else:
-                # TODO Unhardcode this
-                new_prev_img = 'environments/locked.png'
-            new_env = AvatarEnvironment(new_name,
-                                        new_fname,
-                                        new_prev_img,
-                                        new_date,
-                                        new_id,
-                                        new_unlocked,
-                                        new_disp_ord)
-            self._environments[new_name] = new_env
-
-    def _populate_special_category_structures(self, conf_data):
-        """ Populate internal structures related to the special categories
-        such as characters and environments
-        :param conf_data: YAML format configuration structure read from file
-        """
-        special_cat_data = conf_data[self.spec_cat_label]
-
-        for special_cat in self.special_category_labels:
-            active_icon_file = special_cat_data[special_cat]['active_icon']
-            if not os.path.isabs(active_icon_file):
-                active_icon_file = os.path.join(ACTIVE_SPECIAL_CATEGORY_ICONS,
-                                                active_icon_file)
-
-            inactive_icon_file = special_cat_data[special_cat]['inactive_icon']
-            if not os.path.isabs(inactive_icon_file):
-                inactive_icon_file = os.path.join(
-                    INACTIVE_SPECIAL_CATEGORY_ICONS,
-                    inactive_icon_file)
-
-            border_icon_file = special_cat_data[special_cat]['selected_border']
-            if not os.path.isabs(border_icon_file):
-                border_icon_file = os.path.join(PREVIEW_ICONS,
-                                                border_icon_file)
-
-            hover_icon_file = special_cat_data[special_cat]['hover_border']
-            if not os.path.isabs(hover_icon_file):
-                hover_icon_file = os.path.join(PREVIEW_ICONS,
-                                               hover_icon_file)
-
-            self._active_special_category_icons[special_cat] = active_icon_file
-            self._inactive_special_category_icons[special_cat] = \
-                inactive_icon_file
-            self._border_special_cat[special_cat] = border_icon_file
-            self._hover_border_special_cat[special_cat] = hover_icon_file
-            self._special_cat_to_disp_order[special_cat] = \
-                special_cat_data[special_cat]['display_order']
-
-    def get_zindex(self, category):
-        """ Provides the z-index for a specific category
-        :param category: The category whose z-index will be returned
-        :returns: z-index as integer
-        :rtype: integer
-        """
-        if category not in self._cat_to_z_index:
-            logger.warn('Category {} not in available ones'.format(category))
-        else:
-            return self._cat_to_z_index[category]
-
-    def _get_reg_item_cat(self, item_name):
-        """ Get the category of an item if that item belongs to a regular
-        category
-        :param item_name: Item whose category will be returned
-        :returns: category as a string
-        :rtype: string or None
-        """
-        if item_name in self._objects:
-            return self._objects[item_name].category()
-        else:
-            return None
-
-    def get_item_category(self, item_name):
-        """ Get the category of an item/environment if that item/env exists
-        :param item_name: item/env name whose category will be returned
-        :returns: category as a string or None
-        :rtype: string or None
-        """
-        cat_label = self._get_reg_item_cat(item_name)
-        if not cat_label:
-            if item_name not in self._environments:
-                logger.warn('Item {} neither in obj nor in env list'.format(
-                    item_name))
-                return None
-            else:
-                cat_label = self.env_label
-        return cat_label
-
-    def list_available_chars(self):
-        """ Provides a list of available characters
-        :returns: list of characters (list of strings)
-        :rtype: list of stings
-        """
-        return [k for k in self._characters.keys()]
-
-    def _list_all_available_regular_objs(self):
-        """ Provides a list of available regular objects
-        :returns: list of objects (list of strings)
-        :rtype: list of strings
-        """
-        return [k for k in self._objects.keys()]
-
-    def list_all_available_environments(self):
-        """ Provides a list of available environments
-        :returns: list of environments (list of strings)
-        :rtype: list of strings
-        """
-        env_inst_ord = sorted(self._environments.itervalues(),
-                              key=lambda env: env.get_disp_order())
-        return [k.name() for k in env_inst_ord]
-
-    def list_all_available_objs(self):
-        """ Provides a list of available objects
-        :returns: list of objects (list of strings)
-        :rtype: list of strings
-        """
-        ret = self._list_all_available_regular_objs()
-        ret.extend(self.list_all_available_environments())
-        return ret
-
-    def _get_avail_objs_regular_cat(self, category):
-        """ Provides a list of available objects for the specific normal category
-        :returns: [] of object names
-                  None if category is not found
-        :rtype: list of strings or None
-        """
-        if category not in self._object_per_cat:
-            err_msg = "cat {} not found, can't return objects".format(category)
-            logger.warn(err_msg)
-            return None
-        else:
-
-            item_inst_sorted = sorted(self._object_per_cat[category],
-                                      key=lambda obj: obj.get_disp_order())
-            return [it.name() for it in item_inst_sorted]
-
-    def get_avail_objs(self, category):
-        """ Provides a list of available objects for the category provided
-        :returns: [] of object names
-                  None if category is not found
-        :rtype: list of strings or None
-        """
-        if category == self.env_label:
-            return self.list_all_available_environments()
-        else:
-            return self._get_avail_objs_regular_cat(category)
-
-    def _list_available_regular_categories(self):
-        """ Provides a list of available regular categories (not case
-        sensitive)
-        :returns: list of categories (list of strings)
-        :rtype: list of strings
-        """
-        reg_cats_ord = sorted(self._cat_to_disp_order.iteritems(),
-                              key=lambda k: k[1])
-        return [k[0] for k in reg_cats_ord]
-
-    def _list_available_special_categories(self):
-        """ Provides a list of available special categories (not case
-        sensitive)
-        :returns: list of categories (list of strings)
-        :rtype: list of strings
-        """
-        # TODO Temporary hardcoding
-        # return self.special_category_labels
-        # At the time it is only one so it is sorted
-        return [self.env_label]
-
-    def list_available_categories(self):
-        """ Provides a list of available categories (not case sensitive)
-        :returns: list of categories (list of strings)
-        :rtype: list of strings
-        """
-        cats = self._list_available_regular_categories()
-        specs = self._list_available_special_categories()
-
-        cats.extend(specs)
-        return cats
-
-    def get_inactive_category_icon(self, category_name):
-        """ Provides the filename of the inactive icons of the provided
-        category
-        :param category_name: Category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string
-        """
-        icon_path = self._get_inactive_reg_category_icon(category_name) or \
-            self._get_inactive_special_category_icon(category_name)
-
-        if not icon_path:
-            logger.warn(
-                "Cat {} not found, can't provide inactive icon path".format(
-                    category_name)
-            )
-        return icon_path
-
-    def _get_inactive_reg_category_icon(self, category_name):
-        """ Provides the filename of the active icons of the provided
-        regular category
-        :param category_name: Category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        if category_name not in self._inactive_category_icons:
-            return None
-        else:
-            return self._inactive_category_icons[category_name]
-
-    def get_active_category_icon(self, category_name):
-        """ Provides the filename of the active icons of the provided category
-        :param category_name: Category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        icon_path = self._get_active_reg_category_icon(category_name) or \
-            self._get_active_special_category_icon(category_name)
-
-        if not icon_path:
-            logger.warn(
-                "Cat {} not found, can't provide active icon path".format(
-                    category_name)
-            )
-
-        return icon_path
-
-    def _get_active_reg_category_icon(self, category_name):
-        """ Provides the filename of the active icons of the provided active
-        category
-        :param category_name: Category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        if category_name not in self._active_category_icons:
-            return None
-        else:
-            return self._active_category_icons[category_name]
-
-    def get_selected_border(self, category_name):
-        """ Provides the filename of the selected border of the preview icon
-        :param category_name: Category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        icon_path = self._get_selected_border_regular(category_name) or \
-            self._get_selected_border_special(category_name)
-        if not icon_path:
-            logger.warn(
-                "Cat {} not found, can't provide selected border path".format(
-                    category_name)
-            )
-
-        return icon_path
-
-    def _get_selected_border_regular(self, category_name):
-        """ Provides the filename of the selected border of the preview icon
-        :param category_name: Regular category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        if category_name not in self._selected_borders:
-            return None
-        else:
-            return self._selected_borders[category_name]
-
-    def get_hover_border(self, category_name):
-        """ Provides the filename of the hover over border of the preview icon
-        :param category_name: Category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        icon_path = self._get_hover_border_regular(category_name) or \
-            self._get_hover_border_special(category_name)
-
-        if not icon_path:
-            logger.warn(
-                "Cat {} wasn't found, can't provide hover border path".format(
-                    category_name)
-            )
-        return icon_path
-
-    def _get_hover_border_regular(self, category_name):
-        """ Provides the filename of the hover border of the preview icon
-        :param category_name: Regular category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        if category_name not in self._hover_borders:
-            return None
-        else:
-            return self._hover_borders[category_name]
-
-    def _get_inactive_special_category_icon(self, special_category_name):
-        """ Provides the filename of the active icons of the provided category
-        :param special_category_name: Category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        if special_category_name not in self._inactive_special_category_icons:
-            return None
-        else:
-            return self._inactive_special_category_icons[special_category_name]
-
-    def _get_active_special_category_icon(self, special_category_name):
-        """ Provides the filename of the inactive icons of the provided category
-        :param special_category_name: Category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        if special_category_name not in self._active_special_category_icons:
-            return None
-        else:
-            return self._active_special_category_icons[special_category_name]
-
-    def _get_selected_border_special(self, special_category_name):
-        """ Provides the filename of the selected border of the preview icon
-        :param special_category_name: Category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        self._border_special_cat
-        if special_category_name not in self._border_special_cat:
-            return None
-        else:
-            return self._border_special_cat[special_category_name]
-
-    def _get_hover_border_special(self, special_category_name):
-        """ Provides the filename of the hover border of the preview icon
-        :param special_category_name: Category name as a string
-        :returns: path to icon as string or None if category is not found
-        :rtype: string or None
-        """
-        if special_category_name not in self._hover_border_special_cat:
-            return None
-        else:
-            return self._border_special_cat[special_category_name]
-
-    def get_char_preview(self, char_name):
-        """ Provides the preview image for a given character
-        :param char_name: Character whose preview image will be returned
-        :returns: The absolute path to the preview image as a str or
-                  None if the character is not available
-        :rtype: string or None
-        """
-        if char_name not in self._characters:
-            logger.warn(
-                "Char {} not in avail chars, can't return preview img".format(
-                    char_name))
-            return None
-        else:
-            return self._characters[char_name].get_preview_img()
-
-    def get_item_preview(self, item_name):
-        """ Provides the preview image for a given item
-        :param item_name: item whose preview image will be returned
-        :returns: The absolute path to the preview image as a str or
-                  None if the item is not available
-        :rtype: string or None
-        """
-        prev_img = self._get_reg_item_preview(item_name) or \
-            self._get_environment_preview(item_name)
-
-        if not prev_img:
-            logger.warn(
-                "Item {} not in avail objs, can't return preview img".format(
-                    item_name)
-            )
-        return prev_img
-
-    def _get_reg_item_preview(self, item_name):
-        """ Provides the preview image for a given regular item
-        :param item_name: item whose preview image will be returned
-        :returns: The absolute path to the preview image as a str or
-                  None if the item is not available
-        :rtype: string or None
-        """
-        if item_name not in self._objects:
-            return None
-        else:
-            return self._objects[item_name].get_preview_img()
-
-    def _get_environment_preview(self, environment_name):
-        """ Provides the preview image for a given environment
-        :param environment_name: environment whose preview image will be
-                                 returned
-        :returns: The absolute path to the preview image as a str or
-                  None if the environment is not available
-        :rtype: string or None
-        """
-        if environment_name not in self._environments:
-            return None
-        else:
-            return self._environments[environment_name].get_preview_img()
-
-    def is_unlocked(self, item_name):
-        """ Returns whether the item/environment is locked
-        :param item_name: item whose lock status will be returned
-        :returns: True iff it is unlocked
-        :rtype: Boolean or None
-        """
-        lock_state = self._is_unlocked_reg_item(item_name)
-
-        if lock_state is None:
-            lock_state = self._is_unlocked_env(item_name)
-
-        if lock_state is None:
-            logger.warn(
-                "Item {} not in avail objs/envs, can't return lock state".format(
-                    item_name)
-            )
-
-        return lock_state
-
-    def _is_unlocked_reg_item(self, item_name):
-        """ Get lock state of an item which belongs to a regular category
-        :param item_name: Item whose lock state will be returned
-        :retuns: Lock state or None if item doesn't belong to a regular category
-        :rtype: Boolean or None
-        """
-        if item_name not in self._objects:
-            return None
-        else:
-            return self._objects[item_name].is_unlocked()
-
-    def _is_unlocked_env(self, item_name):
-        """ Get lock state of an environment
-        :param item_name: env whose lock state will be returned
-        :retuns: Lock state or None if item isn't an environment
-        :rtype: Boolean or None
-        """
-        if item_name not in self._environments:
-            return None
-        else:
-            return self._environments[item_name].is_unlocked()
 
 
 class AvatarCreator(AvatarConfParser):
